requirements.txt
django
celery
redis
Faker
python-decouple
django-celery-beat 
django-celery-results # currently requires Django 4.0 or older
django-cors-headers
django-filter
django-crispy-forms
django-autocomplete-light
django-allauth
djangorestframework
djangorestframework-simplejwt
django-htmx
django-anymail
django-ckeditor
crispy-bootstrap5
beautifulsoup4
pillow
jupyter
numpy
pandas
matplotlib
scikit-surprise
scikit-learn





